+++ Squidstribution/Assets/Scripts/Enemies/Health/IDamageable.cs
@@ -1,5 +1,5 @@
 public interface IDamageable
 {
     //float CurrentHealth { get; set; }
-    void ApplyDamage(float damage);
+    void ApplyDamage(int damage); 
 }
